Apart from staying up to date, this pulls in https://github.com/typetools/checker-framework/pull/5178 (and follow-up https://github.com/typetools/checker-framework/pull/5210) which should give us a bit of an overhead reduction, particularly for large methods