@msridhar 2022-08-04
Here are the benchmarking numbers.

With this PR:
```
Benchmark                          Mode  Cnt   Score   Error  Units
AutodisposeBenchmark.compile      thrpt   25  10.633 ± 0.049  ops/s
CaffeineBenchmark.compile         thrpt   25   2.497 ± 0.020  ops/s
DFlowMicroBenchmark.compile       thrpt   25  28.560 ± 0.364  ops/s
NullawayReleaseBenchmark.compile  thrpt   25   1.509 ± 0.014  ops/s
```
Before:
```
Benchmark                          Mode  Cnt   Score   Error  Units
AutodisposeBenchmark.compile      thrpt   25  10.643 ± 0.087  ops/s
CaffeineBenchmark.compile         thrpt   25   2.436 ± 0.018  ops/s
DFlowMicroBenchmark.compile       thrpt   25  15.605 ± 0.295  ops/s
NullawayReleaseBenchmark.compile  thrpt   25   1.449 ± 0.014  ops/s
```
There is a huge improvement on the dataflow micro-benchmark (as expected), and a small (but outside the noise) improvement for NullAway itself.  Surprisingly, there seems to be a small overhead regression for Caffeine; I don't really understand why that would happen.  I will see if that repros on another run.